requirements.txt
langchain
openai
streamlit
python-dotenv
bs4
chromadb
tiktoken
streamlit-aggrid